Change stake related types to `u128`

+++ src/mocks.rs
@@ -4,7 +4,7 @@ use crate::validator::RawValidatorData;
 
 /// Generates `num` instances of `RawValidatorData` each having the same `stake`. Out of theses
 /// validators `num_malicious` are malicious.
-pub fn new_validators(num: u64, stake: u64, num_malicious: u64) -> Vec<RawValidatorData> {
+pub fn new_validators(num: u64, stake: u128, num_malicious: u64) -> Vec<RawValidatorData> {
     let mut validators = vec![];
     for i in 0..num {
         let v = RawValidatorData {
